Reorder components folder structure (new folders for Button and Layout components, relocate files from Common to Button and Layout folders)
