+++ app/layout.js
@@ -1,12 +1,12 @@
 // /app/layout.js
 
 import '@/app/globals.css';
-import Navigation from '@/app/components/Common/Navigation';
+import Navigation from '@/app/components/Layout/Navigation';
 import AuthProvider from '@/app/components/Authentication/AuthProvider';
-import Footer from '@/app/components/Common/Footer';
+import Footer from '@/app/components/Layout/Footer';
 import ThemeProvider from '@/app/components/Common/ThemeProvider';
-import PageUpButton from '@/app/components/Common/PageUpButton';
-import MainContainer from '@/app/components/Common/MainContainer';
+import PageUpButton from '@/app/components/Button/PageUpButton';
+import MainContainer from '@/app/components/Layout/MainContainer';
 
 export const metadata = {
   title: '#GMNDR Authentication Demo',
